Working implementation of the Mating Manager
A working implementation of the Mating Manager has been laid. It is,
however, pretty basic and simple. Even though efficiency shouldn't
be a priority --since the couples should be of the order of 100 or
lower-- the current method is clearly inefficient. Since only a
single array of men and women is kept, it is needed that on each
iteration every instance of Man is checked.

In the future, it would be a more efficient aproach to either keep
the married and unmarried men/women in different structures (sets?)
or to keep track of the unmarried men/women, possibly with a
complementary boolean array.